UNTESTED Reduce #ifdef hell on wifi
This fix to the structure of the Wifi code alows for a better link optimisation and so it is no longer necessary for the wifi code or header to read values from the config.h file.

Compilation shows that >11Kb of code is eliminated by the linker if the setup and loop are not called.
~including the interface does not really change anything